change load logic

+++ VRAG_Framework/internVL2_finetuned.py
@@ -10,7 +10,8 @@ from torchvision.transforms.functional import InterpolationMode
 from transformers import AutoModel, AutoTokenizer
 from emb_builder import EmbBuilder
 from emb_module.emb_builder import ClassicEmbBuilder
-from internvl_chat.modeling_internvl_chat import InternVLChatModel
+# from InternVL.internvl_chat.internvl.model.internvl_chat.modeling_internvl_chat import InternVLChatModel
+from internvl.model.internvl_chat.modeling_internvl_chat import InternVLChatModel
 from context_former import ContextFormer
 from utils import split_image, delete_images, merge_dicts, find_longest_diagnosis_keys, expand_disease_abbreviation
 
@@ -31,7 +32,7 @@ class InternVL2_finetuned():
         self.model = InternVLChatModel.from_pretrained(
             args.model_path,
             load_in_8bit=load_8bit,
-            torch_dtype=torch.float16,
+            torch_dtype=torch.bfloat16,
             device_map='auto').eval()
         self.tokenizer = AutoTokenizer.from_pretrained(args.model_path, trust_remote_code=True, use_fast=False)
         # self.model, self.tokenizer = load_model_and_tokenizer(args.model_path)
